requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
pydantic==2.12.0
PyJWT==2.10.1
python-multipart==0.0.20
passlib==1.7.4
psycopg2-binary==2.9.9
pika==1.3.2
pydantic-settings==2.0.3
